refactor image upload/processing script
